Update cpu_simple_sum.hpp
deal with more sxc++ int vs size_t warnings

+++ src/cpu/cpu_simple_sum.hpp
@@ -83,13 +83,13 @@ struct cpu_simple_sum_t: public c_compatible {
             balance211(blocks_number, nthr, ithr, start, end);
 
             for (size_t nb = start; nb < end; ++nb) {
-                int start_e = nb * block_size;
-                int end_e = start_e + block_size;
-                for (int e = start_e; e < end_e; e++) {
+                size_t start_e = nb * block_size;
+                size_t end_e = start_e + block_size;
+                for (size_t e = start_e; e < end_e; e++) {
                     output[e] = static_cast<data_t>(scale_[0] * input_ptrs[0][e]);
                 }
                 for (int a = 1; a < num_arrs; a++) {
-                    for (int e = start_e; e < end_e; e++) {
+                    for (size_t e = start_e; e < end_e; e++) {
                         output[e] += static_cast<data_t>(scale_[a] * input_ptrs[a][e]);
                     }
                 }
@@ -98,11 +98,11 @@ struct cpu_simple_sum_t: public c_compatible {
             if (tail != 0 && ithr == nthr - 1) {
                 size_t start_e = nelems - tail;
                 size_t end_e = nelems;
-                for (int e = start_e; e < end_e; e++) {
+                for (size_t e = start_e; e < end_e; e++) {
                     output[e] = static_cast<data_t>(scale_[0] * input_ptrs[0][e]);
                 }
                 for (int a = 1; a < num_arrs; a++) {
-                    for (int e = start_e; e < end_e; e++) {
+                    for (size_t e = start_e; e < end_e; e++) {
                         output[e] += static_cast<data_t>(scale_[a] * input_ptrs[a][e]);
                     }
                 }
